Extract SimpleSelection tests
